Send emails when a PR is labeled
