+++ functions/src/types.ts
@@ -515,6 +515,7 @@ export namespace github {
     base: Commit;
     _links: Links;
     user: User;
+    labels: Label[];
   }
 }
 
@@ -623,6 +624,39 @@ export namespace internal {
     locked: boolean;
     assignee?: User;
     html_url?: string;
+  }
+
+  export interface IssueOrPullRequest {
+    number: number;
+    title: string;
+    body: string;
+    // TODO: does a PR have these?
+    labels: Label[];
+
+    html_url?: string;
+    // id: number;
+    // url: string;
+    // diff_url: string;
+    // patch_url: string;
+    // issue_url: string;
+    // commits_url: string;
+    // review_comments_url: string;
+    // review_comment_url: string;
+    // comments_url: string;
+    // statuses_url: string;
+    // state: string;
+    // body: string;
+    // assignee: User;
+    // milestone: Milestone;
+    // locked: boolean;
+    // created_at: string;
+    // updated_at: string;
+    // closed_at: string;
+    // merged_at: string;
+    // head: Commit;
+    // base: Commit;
+    // _links: Links;
+    // user: User;
   }
 
   export interface Comment {
